Refactor: mixed agent should be inside __init__.py to prevent circular import
